+++ mapbox-maps-ios/Sources/MapboxMaps/Foundation/Signal.swift
@@ -1,6 +1,9 @@
 /// Signal is a typed interface for observing arbitrary values over time.
 ///
-/// Signal is stateless, which means it doesn't store any payloads sent to it.
+/// Signal delegates observers managing logic to the `observeImpl` closure, but provides flexible interface for observing.
+///
+/// - Note: `Signal` is iOS12-compatible simplified alternative to `Combine.Publisher`. It's behavior is
+/// aligned with Publisher for easier future migration to Combine. If your app supports iOS >= 13.0, use `Signal` as `Combine.Publisher` .
 public struct Signal<Payload> {
     /// Handles received payloads.
     public typealias Handler = (Payload) -> Void
@@ -11,6 +14,8 @@ public struct Signal<Payload> {
     private let observeImpl: ObserveImpl
 
     /// Adds an observer closure that will be called every time signal is triggered.
+    ///
+    /// - Note: Analogous to `sink` in Combine.
     ///
     /// - Parameters:
     ///     - handler: A handler closure.
@@ -31,6 +36,8 @@ public struct Signal<Payload> {
 extension Signal {
     /// Adds an observer closure that will be triggered only once.
     ///
+    /// - Note: Analogous to `prefix(1).sink` in Combine.
+    ///
     /// - Parameters:
     ///     - handler: A handler closure.
     /// - Returns: Cancellable object that is used to cancel the subscription. If it is canceled or deinited the subscription will be cancelled immediately.
@@ -39,6 +46,13 @@ extension Signal {
     }
 }
 
+// NOTE: Signal implements the Combine.Publisher, which means every operator available for
+// Publisher (such as `map`, `prefix`, `sink` and others) may be used on Signal.
+// It means that Signal's naming shouldn't collide with Combine for better user experience.
+// We don't ship full-featured reactive framework (which Combine is) with Maps SDK,
+// so be careful with publishing operators on Signal.
+// Currently only the `observe` and `observeNext` are published which seem to be enough
+// for the most common use cases.
 extension Signal {
     /// Creates a signal that triggers once, then cancels itself.
     internal func takeFirst() -> Signal {